splitting up functions across scripts
