+++ _rmd/extra_loss_moments/utils.py
@@ -2,6 +2,38 @@
 import numpy as np
 import pandas as pd
 from scipy import stats
+from typing import Callable
+from scipy.stats._multivariate import multivariate_normal_frozen
+from scipy.stats._distn_infrastructure import rv_continuous_frozen, rv_discrete_frozen
+
+
+def _is_int(x):
+    """Checks that x is an integer"""
+    return int(x) == x
+
+
+def _input_checks(loss, 
+                dist1, 
+                num_samples = None, 
+                seed = None, 
+                dist2 = None, 
+                k_sd = None
+                ) -> None:
+    """Makes sure that MCI/NumInt arguments are as expected, assertion checks only"""
+    accepted_dists = (multivariate_normal_frozen, rv_continuous_frozen, rv_discrete_frozen)
+    assert isinstance(loss, Callable), f'loss must be a callable function, not {type(loss)}'
+    assert isinstance(dist1, accepted_dists), f'the first distribution must be of type rv_continuous_frozen, not {type(dist1)}'
+    if num_samples is not None:
+        assert num_samples > 1, f'You must have num_samples > 1, not {num_samples}'
+    if seed is not None:
+        assert _is_int(seed), f'seed must be an integer, not {seed}'
+        assert seed >= 0, f'seed must be positive, not {seed}'
+    if dist2 is not None:
+        assert isinstance(dist2, Callable), f'the second distribution must a callable function, not {type(dist2)}'
+    if k_sd is not None:
+        assert k_sd > 0, f'k must be strictkly positive, not {k_sd}'
+
+
 
 expected_dist = stats._multivariate.multivariate_normal_frozen
 def generate_ellipse_points(
